Intermediate commit where things mostly work...
